--- detekt/wrapper/src/main/kotlin/io/buildfoundation/bazel/detekt/Result.kt
@@ -1,11 +1,0 @@
-package io.buildfoundation.bazel.detekt
-
-internal sealed class Result {
-    object Success : Result()
-    data class Failure(val description: String) : Result()
-
-    val consoleStatusCode = when (this) {
-        is Success -> 0
-        is Failure -> 1
-    }
-}
