+++ detekt/wrapper/src/main/kotlin/io/buildfoundation/bazel/detekt/execute/Detekt.kt
@@ -1,9 +1,9 @@
-package io.buildfoundation.bazel.detekt
+package io.buildfoundation.bazel.detekt.execute
 
 import io.gitlab.arturbosch.detekt.cli.buildRunner
 import java.io.PrintStream
 
-internal interface Detekt {
+interface Detekt {
 
     fun execute(args: Array<String>, outputPrinter: PrintStream, errorPrinter: PrintStream)
