+++ detekt/wrapper/src/test/kotlin/io/buildfoundation/bazel/detekt/TestsSuite.kt
@@ -1,5 +1,8 @@
 package io.buildfoundation.bazel.detekt
 
+import io.buildfoundation.bazel.detekt.execute.DetektExecutableTests
+import io.buildfoundation.bazel.detekt.execute.WorkerExecutableTests
+import io.buildfoundation.bazel.detekt.stream.WorkerStreamsTests
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
