--- detekt/wrapper/src/testFixtures/kotlin/io/buildfoundation/bazel/detekt/TestExecutable.kt
@@ -1,9 +1,0 @@
-package io.buildfoundation.bazel.detekt
-
-import io.buildfoundation.bazel.detekt.Executable
-import io.buildfoundation.bazel.detekt.Result
-
-internal class TestExecutable(private val result: Result) : Executable {
-
-    override fun execute(args: Array<String>) = result
-}
