+++ detekt/wrapper/src/testFixtures/kotlin/io/buildfoundation/bazel/detekt/execute/TestExecutable.kt
@@ -1,0 +1,9 @@
+package io.buildfoundation.bazel.detekt.execute
+
+import io.buildfoundation.bazel.detekt.execute.Executable
+import io.buildfoundation.bazel.detekt.execute.Result
+
+internal class TestExecutable(private val result: Result) : Executable {
+
+    override fun execute(args: Array<String>) = result
+}
